split AbstractMeansure
